+++ scripts/embedding/gprolog/gprolog_embed_logtalk.sh
@@ -5,7 +5,7 @@
 ##   This script creates a new GNU Prolog top-level interpreter
 ##   that embeds Logtalk and optionally a Logtalk application
 ## 
-##   Last updated on October 27, 2019
+##   Last updated on March 24, 2022
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   Copyright 1998-2022 Paulo Moura <pmoura@logtalk.org>
@@ -27,7 +27,7 @@
 
 
 print_version() {
-	echo "$(basename "$0") 0.12"
+	echo "$(basename "$0") 0.13"
 	exit 0
 }
 
@@ -99,6 +99,7 @@ fi
 
 # default values
 directory="$(pwd -P)"
+temporary=""
 name="logtalk"
 paths="$LOGTALKHOME/paths/paths_core.pl"
 compile="false"
@@ -111,13 +112,14 @@ usage_help()
 	echo "source code given its loader file."
 	echo
 	echo "Usage:"
-	echo "  $(basename "$0") [-c] [-d directory] [-n name] [-p paths] [-s settings] [-l loader]"
+	echo "  $(basename "$0") [-c] [-d directory] [-t tmpdir] [-n name] [-p paths] [-s settings] [-l loader]"
 	echo "  $(basename "$0") -v"
 	echo "  $(basename "$0") -h"
 	echo
 	echo "Optional arguments:"
 	echo "  -c compile library alias paths in paths and settings files"
 	echo "  -d directory for new top-level (absolute path; default is current directory)"
+	echo "  -t temporary directory for intermediate files (absolute path; default is an auto-created directory)"
 	echo "  -n name of the generated top-level (default is logtalk)"
 	echo "  -p library paths file (absolute path; default is $paths)"
 	echo "  -s settings file (absolute path)"
@@ -128,11 +130,12 @@ usage_help()
 	echo
 }
 
-while getopts "cd:n:p:l:s:vh" option
+while getopts "cd:t:n:p:l:s:vh" option
 do
 	case $option in
 		c) compile="true";;
 		d) d_arg="$OPTARG";;
+		t) t_arg="$OPTARG";;
 		n) n_arg="$OPTARG";;
 		p) p_arg="$OPTARG";;
 		s) s_arg="$OPTARG";;
@@ -148,6 +151,10 @@ args=("$@")
 
 if [ "$d_arg" != "" ] ; then
 	directory="$d_arg"
+fi
+
+if [ "$t_arg" != "" ] ; then
+	temporary="$t_arg"
 fi
 
 if [ "$n_arg" != "" ] ; then
@@ -187,10 +194,12 @@ fi
 
 mkdir -p "$directory"
 
-temporary=$(mktemp -d)
-if [[ ! "$temporary" || ! -d "$temporary" ]]; then
-  echo "Could not create temporary directory!"
-  exit 1
+if [ "$temporary" == "" ] ; then
+	temporary=$(mktemp -d)
+	if [[ ! "$temporary" || ! -d "$temporary" ]]; then
+		echo "Could not create temporary directory!"
+		exit 1
+	fi
 fi
 
 cd "$temporary" || exit 1