+++ scripts/embedding/yap/yap_logtalk_pl.sh
@@ -6,7 +6,7 @@
 ##   and runtime and optionally an application.pl file with a Logtalk
 ##   application
 ## 
-##   Last updated on October 27, 2019
+##   Last updated on March 24, 2022
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   Copyright 1998-2022 Paulo Moura <pmoura@logtalk.org>
@@ -28,7 +28,7 @@
 
 
 print_version() {
-	echo "$(basename "$0") 0.12"
+	echo "$(basename "$0") 0.13"
 	exit 0
 }
 
@@ -102,6 +102,7 @@ fi
 saved_state="false"
 goal="true"
 directory="$(pwd -P)"
+temporary=""
 name="application"
 paths="$LOGTALKHOME/paths/paths_core.pl"
 hooks="$LOGTALKHOME/adapters/yaphooks.pl"
@@ -115,7 +116,7 @@ usage_help()
 	echo "code given its loader file. It can also generate a standalone saved state."
 	echo
 	echo "Usage:"
-	echo "  $(basename "$0") [-c] [-x] [-d directory] [-n name] [-p paths] [-k hooks] [-s settings] [-l loader] [-g goal]"
+	echo "  $(basename "$0") [-c] [-x] [-d directory] [-t tmpdir] [-n name] [-p paths] [-k hooks] [-s settings] [-l loader] [-g goal]"
 	echo "  $(basename "$0") -v"
 	echo "  $(basename "$0") -h"
 	echo
@@ -123,6 +124,7 @@ usage_help()
 	echo "  -c compile library alias paths in paths and settings files"
 	echo "  -x also generate a standalone saved state"
 	echo "  -d directory for generated Prolog files (absolute path; default is current directory)"
+	echo "  -t temporary directory for intermediate files (absolute path; default is an auto-created directory)"
 	echo "  -n name of the generated saved state (default is application)"
 	echo "  -p library paths file (absolute path; default is $paths)"
 	echo "  -k hooks file (absolute path; default is $hooks)"
@@ -134,12 +136,13 @@ usage_help()
 	echo
 }
 
-while getopts "cxd:n:p:k:s:l:g:vh" option
+while getopts "cxd:t:n:p:k:s:l:g:vh" option
 do
 	case $option in
 		c) compile="true";;
 		x) saved_state="true";;
 		d) d_arg="$OPTARG";;
+		t) t_arg="$OPTARG";;
 		n) n_arg="$OPTARG";;
 		p) p_arg="$OPTARG";;
 		k) k_arg="$OPTARG";;
@@ -154,6 +157,10 @@ done
 
 if [ "$d_arg" != "" ] ; then
 	directory="$d_arg"
+fi
+
+if [ "$t_arg" != "" ] ; then
+	temporary="$t_arg"
 fi
 
 if [ "$n_arg" != "" ] ; then
@@ -206,10 +213,12 @@ fi
 
 mkdir -p "$directory"
 
-temporary=$(mktemp -d)
-if [[ ! "$temporary" || ! -d "$temporary" ]]; then
-  echo "Could not create temporary directory!"
-  exit 1
+if [ "$temporary" == "" ] ; then
+	temporary=$(mktemp -d)
+	if [[ ! "$temporary" || ! -d "$temporary" ]]; then
+		echo "Could not create temporary directory!"
+		exit 1
+	fi
 fi
 
 cd "$temporary" || exit 1
@@ -277,7 +286,7 @@ if [ "$loader" != "" ] ; then
 	mkdir -p "$temporary/application"
 	cd "$temporary/application" || exit 1
 	yap -g "consult('$directory/logtalk'),set_logtalk_flag(clean,off),set_logtalk_flag(scratch_directory,'$temporary/application'),logtalk_load('$loader'),halt"
-	cat $(ls -rt "$temporary/application"/*.yap) > "$directory"/application.pl
+	cat $(ls -rt *.yap) > "$directory"/application.pl
 fi
 
 if [ "$saved_state" == "true" ] ; then