+++ RELEASE_NOTES.md
@@ -41,8 +41,8 @@ Installers and installation scripts
 
 * IMPROVED: Embedding examples.
 
-* UPDATED: The SWI-Prolog embedding script to allow specifying the temporary
-directory to be used for intermediate files.
+* UPDATED: Sample embedding scripts to allow specifying the temporary directory
+to be used for intermediate files.
 
 
 3.54.0 - March 22, 2022